refactor(pipeline): decompose engine.go into focused files [#280]
Split the 2778-line engine.go into six focused files by concern:

- engine.go (775 lines): core types, lifecycle (Run/Resume/Confirm),
  phase orchestration (executePhases), pause/checkpoint, timeout, and
  event emission
- engine_phase.go (389 lines): single-phase execution (runPhase),
  retry logic, error classification, backoff, and prompt data building
- engine_review.go (613 lines): parallel-review dispatch, reviewer
  execution with retry, finding merge, verdict computation, and
  review artifact generation
- engine_gate.go (518 lines): domain gating (gatePhase), rework
  routing (gateRework), corrective/patch handling, verify-fail logic,
  and skip-plan-from-triage
- engine_feedback.go (433 lines): rework feedback extraction from
  verify/review results, prior-cycle collection, code snippet
  enrichment, regression detection, and helper utilities
- engine_budget.go (91 lines): pipeline and per-phase/generation
  budget checking with warning thresholds

Pure code motion — no logic changes. All existing tests pass unchanged.
