+++ internal/pipeline/engine_phase.go
@@ -1,0 +1,389 @@
+package pipeline
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"math"
+	"time"
+
+	"github.com/decko/soda/internal/progress"
+	"github.com/decko/soda/internal/runner"
+)
+
+// runPhase executes a single phase including dependency checks, budget checks,
+// worktree creation, prompt rendering, runner invocation with retries, and gating.
+func (e *Engine) runPhase(ctx context.Context, phase PhaseConfig) error {
+	// Polling phases are handled separately.
+	if phase.Type == "polling" {
+		return e.runMonitor(ctx, phase)
+	}
+
+	// Parallel-review phases dispatch multiple reviewers concurrently.
+	if phase.Type == "parallel-review" {
+		return e.runParallelReview(ctx, phase)
+	}
+
+	// Check dependencies.
+	for _, dep := range phase.DependsOn {
+		if !e.state.IsCompleted(dep) {
+			return &DependencyNotMetError{Phase: phase.Name, Dependency: dep}
+		}
+	}
+
+	// Check budget.
+	if err := e.checkBudget(phase); err != nil {
+		return err
+	}
+
+	// Pre-run per-phase budget check: prevent starting a new generation
+	// when cumulative cost already exceeds (or meets) the per-phase limit.
+	if err := e.checkPhaseBudget(phase); err != nil {
+		return err
+	}
+
+	// Mark phase running and notify callback.
+	if err := e.state.MarkRunning(phase.Name); err != nil {
+		return fmt.Errorf("engine: mark running %s: %w", phase.Name, err)
+	}
+	e.emit(Event{Phase: phase.Name, Kind: EventPhaseStarted, Data: map[string]any{"generation": e.state.Meta().Phases[phase.Name].Generation}})
+
+	// Build prompt data and render template.
+	promptData, err := e.buildPromptData(phase)
+	if err != nil {
+		_ = e.state.MarkFailed(phase.Name, err)
+		e.emitPhaseFailed(phase.Name, err)
+		return fmt.Errorf("engine: build prompt data for %s: %w", phase.Name, err)
+	}
+
+	// Inject diff context for corrective phases so the patch prompt
+	// can see what was implemented without reading the plan artifact.
+	if phase.Type == "corrective" {
+		promptData.DiffContext = e.computeDiffContext(ctx)
+	}
+
+	// Store plan hash for staleness guard on phases that depend on plan.
+	for _, dep := range phase.DependsOn {
+		if dep == "plan" {
+			if h := e.computePlanHash(); h != "" {
+				e.state.Meta().Phases[phase.Name].PlanHash = h
+			}
+			break
+		}
+	}
+
+	loadResult, err := e.config.Loader.LoadWithSource(phase.Prompt)
+	if err != nil {
+		_ = e.state.MarkFailed(phase.Name, err)
+		e.emitPhaseFailed(phase.Name, err)
+		return fmt.Errorf("engine: load template for %s: %w", phase.Name, err)
+	}
+
+	// Emit source info so operators can see which template was used.
+	promptEvent := Event{
+		Phase: phase.Name,
+		Kind:  EventPromptLoaded,
+		Data: map[string]any{
+			"source":      loadResult.Source,
+			"is_override": loadResult.IsOverride,
+		},
+	}
+	if loadResult.Fallback {
+		promptEvent.Data["fallback"] = true
+		promptEvent.Data["fallback_reason"] = loadResult.FallbackReason
+	}
+	e.emit(promptEvent)
+
+	rendered, err := RenderPrompt(loadResult.Content, promptData)
+	if err != nil {
+		_ = e.state.MarkFailed(phase.Name, err)
+		e.emitPhaseFailed(phase.Name, err)
+		return fmt.Errorf("engine: render prompt for %s: %w", phase.Name, err)
+	}
+
+	_ = e.state.WriteLog(phase.Name, "prompt", []byte(rendered))
+
+	// Build runner opts. Tighten budget to the smallest remaining limit.
+	remaining := e.config.MaxCostUSD - e.state.Meta().TotalCost
+	if e.config.MaxCostUSD <= 0 {
+		remaining = 0 // no budget enforcement
+	}
+	// Cap with per-phase limit: use the remaining per-phase budget
+	// (MaxCostPerPhase minus cumulative cost already spent) as the tighter bound.
+	if e.config.MaxCostPerPhase > 0 {
+		perPhaseRemaining := e.config.MaxCostPerPhase - e.state.Meta().Phases[phase.Name].CumulativeCost
+		if remaining <= 0 || perPhaseRemaining < remaining {
+			remaining = perPhaseRemaining
+		}
+	}
+	// Cap with per-generation limit: use the remaining generation budget
+	// (MaxCostPerGeneration minus current generation cost) as the tighter bound.
+	if e.config.MaxCostPerGeneration > 0 {
+		genRemaining := e.config.MaxCostPerGeneration - e.state.Meta().Phases[phase.Name].Cost
+		if remaining <= 0 || genRemaining < remaining {
+			remaining = genRemaining
+		}
+	}
+	// Prefer per-phase model if set, otherwise use the global model.
+	model := e.config.Model
+	if phase.Model != "" {
+		model = phase.Model
+	}
+	opts := runner.RunOpts{
+		Phase:        phase.Name,
+		SystemPrompt: rendered,
+		UserPrompt:   "Execute the task described in the system prompt.",
+		OutputSchema: phase.Schema,
+		AllowedTools: phase.Tools,
+		MaxBudgetUSD: remaining,
+		WorkDir:      e.workDir(phase),
+		Model:        model,
+		Timeout:      phase.Timeout.Duration,
+		OnChunk:      e.makeOnChunk(ctx, phase.Name),
+	}
+
+	// Run with retry.
+	result, err := e.runWithRetry(ctx, phase, opts)
+	if err != nil {
+		_ = e.state.MarkFailed(phase.Name, err)
+		e.emitPhaseFailed(phase.Name, err)
+		return err
+	}
+
+	// Record results.
+	if result.Output != nil {
+		if err := e.state.WriteResult(phase.Name, result.Output); err != nil {
+			return fmt.Errorf("engine: write result for %s: %w", phase.Name, err)
+		}
+	}
+	if result.RawText != "" {
+		if err := e.state.WriteArtifact(phase.Name, []byte(result.RawText)); err != nil {
+			return fmt.Errorf("engine: write artifact for %s: %w", phase.Name, err)
+		}
+	}
+	if err := e.state.AccumulateCost(phase.Name, result.CostUSD); err != nil {
+		return fmt.Errorf("engine: accumulate cost for %s: %w", phase.Name, err)
+	}
+
+	// Per-phase cost enforcement: abort if this phase exceeded its budget.
+	if err := e.checkPhaseBudget(phase); err != nil {
+		_ = e.state.MarkFailed(phase.Name, err)
+		e.emitPhaseFailed(phase.Name, err)
+		return err
+	}
+
+	// Mark completed and notify callback.
+	if err := e.state.MarkCompleted(phase.Name); err != nil {
+		return fmt.Errorf("engine: mark completed %s: %w", phase.Name, err)
+	}
+	completedData := map[string]any{
+		"duration_ms": e.state.Meta().Phases[phase.Name].DurationMs,
+		"cost":        e.state.Meta().Phases[phase.Name].Cost,
+	}
+	if result.Output != nil {
+		if summary := progress.PhaseSummary(phase.Name, result.Output); summary != "" {
+			completedData["summary"] = summary
+		}
+	}
+	e.emit(Event{
+		Phase: phase.Name,
+		Kind:  EventPhaseCompleted,
+		Data:  completedData,
+	})
+
+	// Domain gating.
+	return e.gatePhase(phase)
+}
+
+// runWithRetry runs the phase with per-category retry limits.
+func (e *Engine) runWithRetry(ctx context.Context, phase PhaseConfig, opts runner.RunOpts) (*runner.RunResult, error) {
+	remaining := map[string]int{
+		"transient": phase.Retry.Transient,
+		"parse":     phase.Retry.Parse,
+		"semantic":  phase.Retry.Semantic,
+	}
+
+	attempt := 0
+	for {
+		result, err := e.runner.Run(ctx, opts)
+		if err == nil {
+			return result, nil
+		}
+
+		category := classifyError(err)
+
+		left, tracked := remaining[category]
+		if !tracked || left <= 0 {
+			return nil, fmt.Errorf("engine: phase %s failed (%s, no retries left): %w", phase.Name, category, err)
+		}
+		remaining[category]--
+
+		switch category {
+		case "transient":
+			delay := backoff(attempt, e.config.JitterFunc)
+			e.config.SleepFunc(delay)
+			e.emit(Event{
+				Phase: phase.Name,
+				Kind:  EventPhaseRetrying,
+				Data:  map[string]any{"category": category, "attempt": attempt + 1, "delay": delay.String()},
+			})
+
+		case "parse":
+			var pe *runner.ParseError
+			if errors.As(err, &pe) {
+				opts.UserPrompt = opts.UserPrompt + "\n\n[RETRY] Previous attempt failed with parse error: " + pe.Error() + "\nPlease fix the output format."
+			}
+			e.emit(Event{
+				Phase: phase.Name,
+				Kind:  EventPhaseRetrying,
+				Data:  map[string]any{"category": category, "attempt": attempt + 1},
+			})
+
+		case "semantic":
+			var se *runner.SemanticError
+			if errors.As(err, &se) {
+				opts.UserPrompt = opts.UserPrompt + "\n\n[RETRY] Previous attempt returned a semantic error: " + se.Message + "\nPlease address this issue."
+			}
+			e.emit(Event{
+				Phase: phase.Name,
+				Kind:  EventPhaseRetrying,
+				Data:  map[string]any{"category": category, "attempt": attempt + 1},
+			})
+		}
+
+		_ = e.state.WriteLog(phase.Name, fmt.Sprintf("retry_%d", attempt+1),
+			[]byte(fmt.Sprintf("category=%s err=%s", category, err)))
+
+		attempt++
+	}
+}
+
+// classifyError maps an error to a retry category using agent-agnostic
+// runner error types. Backend runners (Claude, sandbox) are responsible
+// for wrapping their specific errors into runner.* types before returning.
+func classifyError(err error) string {
+	if errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded) {
+		return "context"
+	}
+	var te *runner.TransientError
+	if errors.As(err, &te) {
+		return "transient"
+	}
+	var pe *runner.ParseError
+	if errors.As(err, &pe) {
+		return "parse"
+	}
+	var se *runner.SemanticError
+	if errors.As(err, &se) {
+		return "semantic"
+	}
+	return "unknown"
+}
+
+// backoff returns an exponential backoff duration capped at 30s, plus jitter.
+func backoff(attempt int, jitterFunc func(time.Duration) time.Duration) time.Duration {
+	base := 2 * time.Second
+	exp := time.Duration(math.Pow(2, float64(attempt))) * base
+	if exp > 30*time.Second {
+		exp = 30 * time.Second
+	}
+	return exp + jitterFunc(time.Second)
+}
+
+// sleepWithContext runs SleepFunc(d) but returns early if ctx is cancelled.
+// It delegates to e.config.SleepFunc in a goroutine so that test no-op sleeps
+// complete instantly while production sleeps remain cancellable.
+func (e *Engine) sleepWithContext(ctx context.Context, d time.Duration) error {
+	if ctx.Err() != nil {
+		return ctx.Err()
+	}
+	done := make(chan struct{})
+	go func() {
+		e.config.SleepFunc(d)
+		close(done)
+	}()
+	select {
+	case <-done:
+		return nil
+	case <-ctx.Done():
+		return ctx.Err()
+	}
+}
+
+// buildPromptData constructs the PromptData for a phase from its dependencies.
+func (e *Engine) buildPromptData(phase PhaseConfig) (PromptData, error) {
+	data := PromptData{
+		Ticket:        e.config.Ticket,
+		Config:        e.config.PromptConfig,
+		Context:       e.config.PromptContext,
+		DetectedStack: e.config.DetectedStack,
+		WorktreePath:  e.state.Meta().Worktree,
+		Branch:        e.state.Meta().Branch,
+		BaseBranch:    e.config.BaseBranch,
+	}
+
+	for _, dep := range phase.DependsOn {
+		artifact, err := e.state.ReadArtifact(dep)
+		if err != nil {
+			// Not all deps produce artifacts; skip if not found.
+			continue
+		}
+		content := string(artifact)
+
+		switch dep {
+		case "triage":
+			data.Artifacts.Triage = content
+		case "plan":
+			data.Artifacts.Plan = content
+		case "implement":
+			data.Artifacts.Implement = content
+		case "verify":
+			data.Artifacts.Verify = content
+		case "review":
+			data.Artifacts.Review = content
+		case "patch":
+			data.Artifacts.Patch = content
+		case "submit":
+			data.Artifacts.Submit.PRURL = e.extractPRURL()
+		}
+	}
+
+	// Inject rework feedback from configured sources. The FeedbackFrom
+	// list is read from the phase's own config. Sources are tried in
+	// priority order; the first one that produces feedback wins.
+	//
+	// On patch retry (rework cycle), this block re-extracts feedback
+	// from the latest verify/review result on disk — previous feedback
+	// is NOT carried over. Each extractor reads the current result file,
+	// so after verify/review re-run and overwrite their results, the
+	// next rework cycle sees only the new failures. See ReworkFeedback
+	// doc comment for the full reset lifecycle.
+	if sources := phase.FeedbackFrom; len(sources) > 0 {
+		for _, source := range sources {
+			if fb := e.extractFeedbackFrom(source); fb != nil {
+				data.ReworkFeedback = fb
+				eventData := map[string]any{
+					"source":  fb.Source,
+					"verdict": fb.Verdict,
+				}
+				switch fb.Source {
+				case "review":
+					eventData["review_findings"] = len(fb.ReviewFindings)
+				case "verify":
+					eventData["fixes_count"] = len(fb.FixesRequired)
+					eventData["failed_criteria"] = len(fb.FailedCriteria)
+					eventData["code_issues"] = len(fb.CodeIssues)
+					eventData["failed_commands"] = len(fb.FailedCommands)
+				}
+				e.emit(Event{
+					Phase: phase.Name,
+					Kind:  EventReworkFeedbackInjected,
+					Data:  eventData,
+				})
+				break
+			}
+		}
+	}
+
+	return data, nil
+}
